% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master-table.R
\name{append_column_study}
\alias{append_column_study}
\title{Append study column with study name}
\usage{
append_column_study(data, name)
}
\arguments{
\item{data}{Data frame or tibble.}

\item{name}{Name of the study.}
}
\value{
Tibble with the data and appended study column.
}
\description{
Append column named 'study' to data, with all values as the
  study name. Does not change the data if a 'study' column already exists.
}
\examples{
dat <- data.frame(
  foo = c("bar0", "bar1", "bar2"),
  bar = c("foo0", "foo1", "foo2")
)
study_dat <- append_column_study(dat, "cool_study")
# Now has study column so cannot change study column
# value with this function
study_dat <- append_column_study(study_dat, "neato_study")
}
