% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master-table.R
\name{create_study_table}
\alias{create_study_table}
\title{Create the master table}
\usage{
create_study_table(study_view, study_name, keys)
}
\arguments{
\item{study_view}{Data frame with metadata file information, including
column names 'id' and 'name'.}

\item{study_name}{Name of the study.}

\item{keys}{The names of the columns that are wanted in the table.}
}
\value{
Data frame with master table data.
}
\description{
Create the master table with the given keys as columns.
}
