% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dict_to_list}
\alias{dict_to_list}
\title{Convert Synapse annotation dictionary to a list}
\usage{
dict_to_list(annotations)
}
\arguments{
\item{annotations}{Annotations dictionary from Synapse.}
}
\value{
List of annotations.
}
\description{
Convert Synapse annotation dictionary to a list.
}
