% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master-table.R
\name{generate_model_master_table}
\alias{generate_model_master_table}
\title{Generate master table for model}
\usage{
generate_model_master_table(
  parent,
  keys = c("individualID", "specimenID", "assay", "species")
)
}
\arguments{
\item{parent}{Parent synID that holds study folders for a model.}

\item{keys}{The names of the columns that are wanted in the table.}
}
\value{
The master data table.
}
\description{
Generate master table for model based on metadata files.
Synapse directory structure is expected to be:
  parent_folder
    |_study_folder
    |   |_ metadata_folder
    |        |_ metadata_file_1
    |        |_ metadata_file_2
    |_study_folder
        |_ metadata_folder
             |_ metadata_file_3
}
