% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master-table.R
\name{get_all_metadata_folder_info}
\alias{get_all_metadata_folder_info}
\title{Get all metadata folder information}
\usage{
get_all_metadata_folder_info(parent)
}
\arguments{
\item{parent}{synID for the parent folder.}
}
\value{
Dataframe with the metadata folder information.
}
\description{
Get all metadata folder information.
 Synapse directory structure is expected to be:
  parent_folder
    |_study_folder
    |   |_ metadata_folder
    |        |_ metadata_file_1
    |        |_ metadata_file_2
    |_study_folder
        |_ metadata_folder
             |_ metadata_file_3
}
