% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse-directory-structure.R
\name{get_children}
\alias{get_children}
\title{Get children of Synapse folder}
\usage{
get_children(parent)
}
\arguments{
\item{parent}{The synID of the parent folder.}
}
\value{
Synapse entity list with child folders/files.
}
\description{
Get a list of the child folders/files within a Synapse parent
folder.
}
\examples{
\dontrun{
library(synapser)
synLogin()
get_children("syn123456")
}
}
