% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_file_data}
\alias{get_file_data}
\title{Get data from file}
\usage{
get_file_data(path)
}
\arguments{
\item{path}{File path}
}
\value{
A data frame with the file data
}
\description{
Get data from file. File types handled are:
  - csv
  - tsv
  - txt (assumed to be tab delimited)
  - xlsx (assumed to be single sheet)
  - xls (assumed to be single sheet)
}
