% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master-table.R
\name{get_metadata_file_and_annotations}
\alias{get_metadata_file_and_annotations}
\title{Get data and annotations from metadata file}
\usage{
get_metadata_file_and_annotations(file_id, file_name, keys)
}
\arguments{
\item{file_id}{Synapse ID for the metadata file.}

\item{file_name}{Name of metadata file.}

\item{keys}{The names of the columns that are wanted in the table.}
}
\value{
Named list where `metadata` refers to a dataframe with the file's
  metadata, and `annotations` refers to a tibble with the file's non-null
  annotations
}
\description{
Get data and annotations from metadata file.
}
