% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master-table.R
\name{get_metadata_file_info}
\alias{get_metadata_file_info}
\title{Get metadata file information}
\usage{
get_metadata_file_info(folder_id, study_name)
}
\arguments{
\item{folder_id}{synID for the metadata folder.}

\item{study_name}{Name of the study.}
}
\value{
Dataframe with all the metadata file information.
}
\description{
Get metadata file information. Assumes the directory structure
looks like:
   metadata_folder
     |_ metadata_file_1
     |_ metadata_file_2
}
