% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master-table.R
\name{get_study_metadata_folder_info}
\alias{get_study_metadata_folder_info}
\title{Get all metadata folder information helper}
\usage{
get_study_metadata_folder_info(study, study_name)
}
\arguments{
\item{study}{The synID for the study folder.}

\item{study_name}{Name of the study.}
}
\value{
Dataframe with the metadata folder information.
}
\description{
Get study metadata folder information. Helper function for
[get_all_metadata_folder_info()]. Assumes the directory structure looks like:
    study_folder
      |_ metadata_folder
          |_ metadata_file_1
          |_ metadata_file_2
}
