% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_into_synapse}
\alias{log_into_synapse}
\title{Log into Synapse}
\usage{
log_into_synapse(...)
}
\arguments{
\item{...}{Additional parameters sent on to [synapser::synLogin()]}
}
\description{
Logs user into Synapse.
}
