% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse-directory-structure.R
\name{move_folder_file}
\alias{move_folder_file}
\title{Move a Synapse folder or file to new parent folder}
\usage{
move_folder_file(id, parent_id)
}
\arguments{
\item{id}{synID of the folder or file to move.}

\item{parent_id}{synID of the new folder to move the folder or file to.}
}
\value{
The Synapse entity that was stored in a new location.
}
\description{
Move a Synapse folder or file to new parent folder.
}
