% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tables.R
\name{remove_na_rows}
\alias{remove_na_rows}
\title{Remove NA rows}
\usage{
remove_na_rows(data, columns = NULL)
}
\arguments{
\item{data}{Data frame with metadata}

\item{columns}{The columns to compare for `NA` values.}
}
\value{
The data without rows that are `NA` across the given columns.
}
\description{
Removes NA rows. If columns is not `NULL`, then will only remove
rows in which all columns given have `NA`. If columns is `NULL`, then will
remove rows that are `NA` across all columns.
}
\examples{
\dontrun{
dat <- data.frame(
  foo = c("bar0", NA, NA, "bar3"),
  bar = c(NA, "foo1", NA, "foo3")
)
dat1 <- remove_na_rows(dat)
dat2 <- remove_na_rows(dat, "foo")
dat3 <- remove_na_rows(dat, c("foo", "bar"))
}
}
