% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tables.R
\name{save_table_to_synapse}
\alias{save_table_to_synapse}
\title{Save table to Synapse}
\usage{
save_table_to_synapse(
  data,
  table_id = NULL,
  project_id = NULL,
  name = NULL,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{data}{Data frame with metadata}

\item{table_id}{synID for the table. If `NULL`, assumes the table needs to be
created first.}

\item{project_id}{synID for the project the table should be uploaded to, if
table needs to be created.}

\item{name}{Desired table name, if table needs to be created.}

\item{overwrite}{If overwrite is `TRUE`, the table contents will be deleted
and the new data uploaded in its place. Note that the schema will not
be updated so the new data must follow the same schema. If `FALSE`, the
new data will be joined to the current table.}

\item{...}{Additional parameters sent to [synapser::synStore()]}
}
\value{
Synapse table entity
}
\description{
Save table to Synapse.
}
